Aula 01, 02 e 05 + Checkpoint I

+++ bimestre-03/programacao-orientada-objetos/modulo01-introducao/aula01-java/aula01-intelliJ/src/Main.java
@@ -3,11 +3,12 @@ public class Main {
     public static void main(String[] args) {
         int valor1 = 15;
         int valor2 = 2;
-        int resultado = valor1 / valor2;
+        float resultado = (float) valor1 / valor2;
+
 
         System.out.println("Hello, world! Resultado: " + resultado);
         //Objeto "System" > subclasse "out" > método "println"
     }
 }
 
-// Atalho Run: Ctrl + shift + F10
+//Atalho Run: Ctrl + shift + F10
